pyproject.toml
[tool.poetry]
name = "Meloetta"
version = "2.0.0a"
description = ""
authors = ["Joshua <josh@josh-is.gay>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8.0"
"discord.py" = "^1.6.0"
discord-ext-menus = {git = "https://github.com/rapptz/discord-ext-menus.git"}
psutil = "^5.7.2"
PyYAML = "^5.3.1"
wavelink = "^0.9.6"
jishaku = "^1.20.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
